+++ datasources/bampfa/solrETL-public.sh
@@ -17,8 +17,9 @@ mv 4solr.*.csv.gz /tmp
 # eases maintainance. ergo, the TENANT parameter
 ##############################################################################
 TENANT=$1
+CORE=public
 SERVER="dba-postgres-prod-42.ist.berkeley.edu port=5313 sslmode=prefer"
-USERNAME="reporter_$TENANT"
+USERNAME="reporter_${TENANT}"
 DATABASE="${TENANT}_domain_${TENANT}"
 CONNECTSTRING="host=$SERVER dbname=$DATABASE"
 ##############################################################################
@@ -38,7 +39,7 @@ time perl -pe 's/[\r\n]/ /g;s/\@\@/\n/g' m1.csv > media.csv
 time psql -R"@@" -F $'\t' -A -U $USERNAME -d "$CONNECTSTRING" -f blobs.sql -o b1.csv
 time perl -pe 's/[\r\n]/ /g;s/\@\@/\n/g' b1.csv > blobs.csv
 # Compute the "view status" of each object
-time perl addStatus.pl public 37 38 < d4.csv > metadata.csv
+time perl addStatus.pl ${CORE} 37 38 < d4.csv > metadata.csv
 # make the header
 head -1 metadata.csv > header4Solr.csv
 # add the blob field name to the header (the header already ends with a tab); rewrite objectcsid_s to id (for solr id...)
@@ -51,28 +52,42 @@ cat header4Solr.csv d7.csv > d8.csv
 ##############################################################################
 # compute _i values for _dt values (to support BL date range searching)
 ##############################################################################
-time python3 computeTimeIntegers.py d8.csv 4solr.$TENANT.public.csv
+time python3 computeTimeIntegers.py d8.csv 4solr.${TENANT}.${CORE}.csv
 wc -l *.csv
-# clear out the existing data
-curl -S -s "http://localhost:8983/solr/${TENANT}-public/update" --data '<delete><query>*:*</query></delete>' -H 'Content-type:text/xml; charset=utf-8'
-curl -S -s "http://localhost:8983/solr/${TENANT}-public/update" --data '<commit/>' -H 'Content-type:text/xml; charset=utf-8'
+##############################################################################
+# check if we have enough data to be worth refreshing...
+##############################################################################
+CSVFILE="4solr.${TENANT}.${CORE}.csv"
+# this value is an approximate lower bound on the number of rows there should
+# be, based on data as of 2019-09-11. It may need to be periodically adjusted.
+MINIMUM=22000
+ROWS=`wc -l < ${CSVFILE}`
+if (( ${ROWS} < ${MINIMUM} )); then
+   echo "Only ${ROWS} rows in ${CSVFILE}; refresh aborted, core left untouched." | mail -s "PROBLEM with ${TENANT}-${CORE} nightly solr refresh" -- cspace-support@lists.berkeley.edu
+   exit 1
+fi
+##############################################################################
+# OK, we are good to go! clear out the existing data
+##############################################################################
+curl -S -s "http://localhost:8983/solr/${TENANT}-${CORE}/update" --data '<delete><query>*:*</query></delete>' -H 'Content-type:text/xml; charset=utf-8'
+curl -S -s "http://localhost:8983/solr/${TENANT}-${CORE}/update" --data '<commit/>' -H 'Content-type:text/xml; charset=utf-8'
 # note: current location, current crate, appraised values have all been redacted
 # in the sql queries themselves.
 # some values were needed for computing the status field (i.e. "on view")
 # TODO however we could also skip them in the Solr load as well...
-time curl -X POST -S -s "http://localhost:8983/solr/${TENANT}-public/update/csv?commit=true&header=true&trim=true&separator=%09&f.grouptitle_ss.split=true&f.grouptitle_ss.separator=;&f.othernumbers_ss.split=true&f.othernumbers_ss.separator=;&f.blob_ss.split=true&f.blob_ss.separator=,&encapsulator=\\" -T 4solr.$TENANT.public.csv -H 'Content-type:text/plain; charset=utf-8' &
+time curl -X POST -S -s "http://localhost:8983/solr/${TENANT}-${CORE}/update/csv?commit=true&header=true&trim=true&separator=%09&f.grouptitle_ss.split=true&f.grouptitle_ss.separator=;&f.othernumbers_ss.split=true&f.othernumbers_ss.separator=;&f.blob_ss.split=true&f.blob_ss.separator=,&encapsulator=\\" -T 4solr.${TENANT}.${CORE}.csv -H 'Content-type:text/plain; charset=utf-8' &
 ##############################################################################
 # count the types and tokens in the final file, check cell counts
 ##############################################################################
-time python3 evaluate.py 4solr.$TENANT.public.csv /dev/null > counts.public.csv &
+time python3 evaluate.py 4solr.${TENANT}.${CORE}.csv /dev/null > counts.${CORE}.csv &
 # get rid of intermediate files
 rm d?.csv m?.csv b?.csv media.csv metadata.csv &
-cut -f43 4solr.${TENANT}.public.csv | grep -v 'blob_ss' |perl -pe 's/\r//' |  grep . | wc -l > counts.public.blobs.csv
-cut -f43 4solr.${TENANT}.public.csv | perl -pe 's/\r//;s/,/\n/g;s/\|/\n/g;' | grep -v 'blob_ss' | grep . | wc -l >> counts.public.blobs.csv
-cp counts.public.blobs.csv /tmp/$TENANT.counts.public.blobs.csv
-cat counts.public.blobs.csv
+cut -f43 4solr.${TENANT}.${CORE}.csv | grep -v 'blob_ss' |perl -pe 's/\r//' |  grep . | wc -l > counts.${CORE}.blobs.csv
+cut -f43 4solr.${TENANT}.${CORE}.csv | perl -pe 's/\r//;s/,/\n/g;s/\|/\n/g;' | grep -v 'blob_ss' | grep . | wc -l >> counts.${CORE}.blobs.csv
+cp counts.${CORE}.blobs.csv /tmp/${TENANT}.counts.${CORE}.blobs.csv
+cat counts.${CORE}.blobs.csv
 wait
-cp counts.public.csv /tmp/$TENANT.counts.public.csv
+cp counts.${CORE}.csv /tmp/${TENANT}.counts.${CORE}.csv
 # zip up .csvs, save a bit of space on backups
 gzip -f *.csv
 date
